utils/images/requirements.txt
gitdb==4.0.12
GitPython==3.1.45
pillow==12.0.0
smmap==5.0.2
